+++ frontend/rust-lib/flowy-chat/src/local_ai/watch.rs
@@ -1,4 +1,5 @@
 use crate::local_ai::local_llm_resource::WatchDiskEvent;
+use crate::local_ai::path::{install_path, offline_app_path};
 use flowy_error::{FlowyError, FlowyResult};
 use notify::{Event, RecursiveMode, Watcher};
 use std::path::PathBuf;
@@ -9,29 +10,6 @@ pub struct WatchContext {
   #[allow(dead_code)]
   watcher: notify::RecommendedWatcher,
   pub path: PathBuf,
-}
-
-pub(crate) fn install_path() -> Option<PathBuf> {
-  #[cfg(target_os = "windows")]
-  return None;
-
-  #[cfg(target_os = "macos")]
-  return Some(PathBuf::from("/usr/local/bin"));
-
-  #[cfg(target_os = "linux")]
-  return None;
-}
-
-pub(crate) fn offline_app_path() -> PathBuf {
-  let offline_app = "appflowy_ai_plugin";
-  #[cfg(target_os = "windows")]
-  return PathBuf::from(format!("/usr/local/bin/{}", offline_app));
-
-  #[cfg(target_os = "macos")]
-  return PathBuf::from(format!("/usr/local/bin/{}", offline_app));
-
-  #[cfg(target_os = "linux")]
-  return PathBuf::from(format!("/usr/local/bin/{}", offline_app));
 }
 
 pub fn watch_offline_app() -> FlowyResult<(WatchContext, UnboundedReceiver<WatchDiskEvent>)> {
